requirements.txt
flask
pyzbar
Pillow
